pyproject.toml
[project]
name = "db-upload"
version = "0.1.0"
description = "A repository for taking csv or json files and uploading them to a postgresql database"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psycopg[binary]>=3.2.0",   # PostgreSQL driver
    "pandas>=2.2.0",             # CSV/JSON parsing and data manipulation
    "sqlalchemy>=2.0.0",         # ORM / engine for DB interactions
    "python-dotenv>=1.0.0",      # Load connection config from .env
    "typer>=0.12.0",             # CLI interface
    "rich>=13.0.0",              # Pretty terminal output
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pandas-stubs>=2.2.0",
]

[project.scripts]
db_upload = "db_upload.cli:app"

[project.urls]
Repository = "https://github.com/your-org/db-upload"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/db_upload"]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=db_upload --cov-report=term-missing"